requirements.txt
# requirements.txt
selenium==3.141.0
python-docx==0.8.10
comtypes==1.1.10


